Allow 1-3 doors and fix while loop crash
The transition animation initially did not work
when there was a single door, so I added an
additional timeout() to let everything catch up.

The site would also crash in the event of a
duplicate result, as the while loop would nver
resolve (no idea how I didn't catch it sooner),
so now the code correctly exits when a good result
is found.